+++ SerialCommunication/source/ModbusRTU.cpp
@@ -1,4 +1,4 @@
-#include "ModbusRTU.h"
+#include "../include/ModbusRTU.h"
 
 ModbusRTU::ModbusRTU()
 {
