+++ SerialCommunication/source/UART.cpp
@@ -1,5 +1,5 @@
-#include "UART.h"
-#include "Predefined.h"
+#include "../include/UART.h"
+#include "../include/Predefined.h"
 
 bool UART::open()
 {
