+++ SerialCommunication/source/main.cpp
@@ -1,5 +1,5 @@
-#include "UART.h"
-#include "ModbusRTU.h"
+#include "../include/UART.h"
+#include "../include/ModbusRTU.h"
 #include <windows.h>
 
 int main()
