+++ SocketCommunication/include/ClientHSMS.h
