+++ SocketCommunication/include/ClientTCP.h
