+++ SocketCommunication/include/Predefined.h
