+++ SocketCommunication/include/ServerHSMS.h
