+++ SocketCommunication/include/ServerHTTP.h
