+++ SocketCommunication/source/ClientHTTP.cpp
@@ -1,4 +1,4 @@
-#include "ClientHTTP.h"
+#include "../include/ClientHTTP.h"
 
 HTTP_RESPONSE ClientHTTP::parseResponseMessage(std::string _msg)
 {
