+++ SocketCommunication/source/ServerHTTP.cpp
@@ -1,4 +1,4 @@
-#include "ServerHTTP.h"
+#include "../include/ServerHTTP.h"
 
 HTTP_REQUEST ServerHTTP::parseRequestMessage(std::string _msg)
 {
