+++ SocketCommunication/source/UDP.cpp
@@ -1,5 +1,5 @@
-#include "UDP.h"
-#include "Predefined.h"
+#include "../include/UDP.h"
+#include "../include/Predefined.h"
 
 UDP::UDP()
 {
