+++ SocketCommunication/source/main.cpp
@@ -1,5 +1,5 @@
-#include "ServerHTTP.h"
-#include "ClientHTTP.h"
+#include "../include/ServerHTTP.h"
+#include "../include/ClientHTTP.h"
 
 int main()
 {
